Moved ResultViewTrackModel outside of ResultViewTrack
