+++ px-checkout/src/main/java/com/mercadopago/android/px/tracking/internal/views/ResultViewTrack.java
@@ -1,13 +1,9 @@
 package com.mercadopago.android.px.tracking.internal.views;
 
 import android.support.annotation.NonNull;
-import com.mercadopago.android.px.internal.util.PaymentDataHelper;
 import com.mercadopago.android.px.model.PaymentResult;
 import com.mercadopago.android.px.preferences.CheckoutPreference;
-import com.mercadopago.android.px.tracking.internal.mapper.FromPaymentMethodToAvailableMethods;
-import com.mercadopago.android.px.tracking.internal.model.AvailableMethod;
-import com.mercadopago.android.px.tracking.internal.model.TrackingMapModel;
-import java.math.BigDecimal;
+import com.mercadopago.android.px.tracking.internal.model.ResultViewTrackModel;
 import java.util.Locale;
 import java.util.Map;
 
@@ -66,39 +62,5 @@ public class ResultViewTrack extends ViewTracker {
     @Override
     public String getViewPath() {
         return String.format(Locale.US, PATH, getMappedResult(payment));
-    }
-
-    private static final class ResultViewTrackModel extends TrackingMapModel {
-
-        private final String style;
-        private final Long paymentId;
-        private final String paymentStatus;
-        private final String paymentStatusDetail;
-        private final String currencyId;
-        private final boolean hasSplitPayment;
-        private final BigDecimal preferenceAmount;
-        private final BigDecimal discountCouponAmount;
-        private AvailableMethod availableMethod;
-
-        ResultViewTrackModel(@NonNull final Style style, @NonNull final PaymentResult payment,
-            @NonNull final CheckoutPreference checkoutPreference) {
-            this.style = style.value;
-            paymentId = payment.getPaymentId();
-            paymentStatus = payment.getPaymentStatus();
-            paymentStatusDetail = payment.getPaymentStatusDetail();
-            currencyId = checkoutPreference.getSite().getCurrencyId();
-            hasSplitPayment = PaymentDataHelper.isSplitPayment(payment.getPaymentDataList());
-            preferenceAmount = checkoutPreference.getTotalAmount();
-            discountCouponAmount = PaymentDataHelper.getTotalDiscountAmount(payment.getPaymentDataList());
-
-            if (payment.getPaymentData() != null && payment.getPaymentData().getPaymentMethod() != null) {
-                availableMethod =
-                    new FromPaymentMethodToAvailableMethods().map(payment.getPaymentData().getPaymentMethod());
-            }
-        }
-
-        public Long getPaymentId() {
-            return paymentId;
-        }
     }
 }
